Added Game Support Demo
This commit is a demo for the future game support of the program. Adding a single game along with an updated ActiveTrial.py file that has a button to launch the game. For the game to be controlled by the exoskeleton, a virtualController.py script was added. GUI.py creates a new virtualController object and refrences realTimeProcessor. The object then takes the incomming data from realTimeProcessor and converts it to gamepad input for the game to read.